Same level for components Audience and Users recipients.
